Conflicts:
	main/src/album/cuesheet.cpp
	main/src/album/track.cpp
	main/src/album/track.hpp
	main/src/widget/player.cpp

+++ main/src/album/index.hpp
@@ -76,7 +76,7 @@ namespace khopper {
 			/**
 			 * @brief Convert to millisecond
 			 */
-			int toMillisecond() const;
+			int64_t toMillisecond() const;
 			/**
 			 * @brief Convert to second
 			 */
@@ -91,7 +91,7 @@ namespace khopper {
 			 * @brief Convert from int
 			 * @param [in] millisecond Time stamp in millisecond
 			 */
-			static Index fromMillisecond( int millisecond );
+			static Index fromMillisecond( int64_t millisecond );
 			/**
 			 * @brief Convert from double
 			 * @param [in] second Time stamp in second
